# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 电气车间备件管理系统
  version: 1.0.0
paths:
...
  /api/v1/measurement-units:
    get:
      tags:
      - 基础数据
      summary: Units
      operationId: units_api_v1_measurement_units_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_MeasurementUnitRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - 基础数据
      summary: Add Unit
      operationId: add_unit_api_v1_measurement_units_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementUnitCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementUnitRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/measurement-units/{item_id}:
    patch:
      tags:
      - 基础数据
      summary: Edit Unit
      operationId: edit_unit_api_v1_measurement_units__item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementUnitUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementUnitRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/project-subitems:
    get:
      tags:
      - 基础数据
      summary: Projects
      operationId: projects_api_v1_project_subitems_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectSubitemRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - 基础数据
      summary: Add Project
      operationId: add_project_api_v1_project_subitems_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSubitemCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSubitemRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/project-subitems/{item_id}:
    patch:
      tags:
      - 基础数据
      summary: Edit Project
      operationId: edit_project_api_v1_project_subitems__item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSubitemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSubitemRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/users:
    get:
      tags:
      - 基础数据
      summary: Users
      operationId: users_api_v1_users_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_UserRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - 基础数据
      summary: Add User
      operationId: add_user_api_v1_users_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/users/{item_id}:
    patch:
      tags:
      - 基础数据
      summary: Edit User
      operationId: edit_user_api_v1_users__item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - code
      - message
      - request_id
      title: ApiError
...
    MeasurementUnitCreate:
      properties:
        code:
          type: string
          maxLength: 32
          minLength: 1
          title: Code
        name:
          type: string
          maxLength: 32
          minLength: 1
          title: Name
        decimal_places:
          type: integer
          maximum: 3.0
          minimum: 0.0
          title: Decimal Places
          default: 0
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - code
      - name
      title: MeasurementUnitCreate
    MeasurementUnitRead:
      properties:
        id:
          type: integer
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        decimal_places:
          type: integer
          title: Decimal Places
        enabled:
          type: boolean
          title: Enabled
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - code
      - name
      - decimal_places
      - enabled
      - version
      title: MeasurementUnitRead
    MeasurementUnitUpdate:
      properties:
        code:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 1
          - type: 'null'
          title: Code
        name:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 1
          - type: 'null'
          title: Name
        decimal_places:
          anyOf:
          - type: integer
            maximum: 3.0
            minimum: 0.0
          - type: 'null'
          title: Decimal Places
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
        version:
          type: integer
          title: Version
      additionalProperties: false
      type: object
      required:
      - version
      title: MeasurementUnitUpdate
...
    Page_MeasurementUnitRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MeasurementUnitRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[MeasurementUnitRead]
    Page_ProjectSubitemRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectSubitemRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[ProjectSubitemRead]
...
    Page_UserRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[UserRead]
...
    ProjectSubitemCreate:
      properties:
        project_code:
          type: string
          maxLength: 64
          minLength: 1
          title: Project Code
        project_name:
          type: string
          maxLength: 128
          minLength: 1
          title: Project Name
        subitem_no:
          type: string
          maxLength: 64
          minLength: 1
          title: Subitem No
        subitem_name:
          type: string
          maxLength: 128
          minLength: 1
          title: Subitem Name
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - project_code
      - project_name
      - subitem_no
      - subitem_name
      title: ProjectSubitemCreate
    ProjectSubitemRead:
      properties:
        id:
          type: integer
          title: Id
        project_code:
          type: string
          title: Project Code
        project_name:
          type: string
          title: Project Name
        subitem_no:
          type: string
          title: Subitem No
        subitem_name:
          type: string
          title: Subitem Name
        enabled:
          type: boolean
          title: Enabled
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - project_code
      - project_name
      - subitem_no
      - subitem_name
      - enabled
      - version
      title: ProjectSubitemRead
    ProjectSubitemUpdate:
      properties:
        project_code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Project Code
        project_name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Project Name
        subitem_no:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Subitem No
        subitem_name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Subitem Name
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
        version:
          type: integer
          title: Version
      additionalProperties: false
      type: object
      required:
      - version
      title: ProjectSubitemUpdate
...
    Role:
      type: string
      enum:
      - SUPER_ADMIN
      - WAREHOUSE_ADMIN
      - PURCHASE_ADMIN
      - READ_ONLY
      title: Role
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
        display_name:
          type: string
          maxLength: 128
          minLength: 1
          title: Display Name
        role:
          $ref: '#/components/schemas/Role'
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - display_name
      - role
      title: UserCreate
    UserRead:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        role:
          $ref: '#/components/schemas/Role'
        enabled:
          type: boolean
          title: Enabled
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - username
      - display_name
      - role
      - enabled
      - version
      title: UserRead
    UserUpdate:
      properties:
        display_name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Display Name
        password:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 6
          - type: 'null'
          title: Password
        role:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
        version:
          type: integer
          title: Version
      additionalProperties: false
      type: object
      required:
      - version
      title: UserUpdate
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer